backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
pydantic
python-dotenv
sentence-transformers
faiss-cpu
beautifulsoup4
requests
reportlab
